a few more loop-related fixes

+++ tests/mem.cpp
@@ -10,7 +10,7 @@ TEST_BOTH(01_gather) {
 }
 
 TEST_BOTH(02_gather_mask) {
-    Mask r = eq(arange<Int32>(100) & 1, 1);
+    Mask r = eq(arange<Int32>(100) & Int32(1), 1);
     UInt32 index = UInt32(33, 62, 75, 2);
     Mask ref = UInt32(1, 0, 1, 0);
     Mask value = gather<Mask>(r, index);
@@ -27,7 +27,7 @@ TEST_BOTH(03_gather_masked) {
 }
 
 TEST_BOTH(04_gather_mask_masked) {
-    Mask r = eq(arange<Int32>(100) & 1, 1);
+    Mask r = eq(arange<Int32>(100) & Int32(1), 1);
     UInt32 index = UInt32(33, 62, 75, 2);
     Mask ref = UInt32(0, 0, 1, 0);
     Mask mask = index > 50;
